requirements.txt
Flask==2.3.3
boto3==1.28.25
python-dotenv==1.0.0